Update frameBuffer.hpp
Enhance FrameBuffer to manage frames with indices

Introduces a new `Frame` structure to encapsulate `cv::Mat` and frame index.
Changes the internal buffer of `FrameBuffer` to store `Frame` objects,
adds a `frameIdx_` member for tracking the current frame index,
and updates the constructor to set a default max size of 10.
Modifies the `pop` method to return a `Frame` and adds
a new method `getFrameByIndex` for frame retrieval by index.